refactor: Extract ClipboardIconButton
