+++ apps/web/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './quick-start';
 export * from './IconButton';
+export * from './ClipboardIconButton';
